requirements.txt
Django==4.1.7
psycopg2-binary==2.9.5
crispy-bootstrap5==0.7
Pillow==9.4.0
pytest==7.2.2
pytest-django==4.5.2
djangorestframework==3.14.0
scrapy==2.8.0
redis==4.5.4
pytest-django>=4.5.2
factory-boy>=3.2.1
pytest-factoryboy>=2.5.0
faker>=13.15.0